Add files via upload
When learning a new language (either foreign or programming language) it is very important to practice for at least 100 hours to feel proficient and comfortable with the language and your ability to code on it. 
So, this program prompts the user to start practicing a new language, calculates how long the practice lasted, and what was accomplished on a 100-hour scale.
The program also asks the user if she/he wants to keep or dismiss the record of the practice to prevent fail positive (the case where the user forgot to stop the timer or wants to dismiss the latest practice).